fix file read()/write() return values, add fstat support
git-svn-id: svn+ssh://market.scs.stanford.edu/home/u1/asbestos/repo/jos64/trunk@1145 927ebb28-5305-0410-b3ca-c4b640df69b7

+++ lib/netd_fd.c
@@ -133,7 +133,7 @@ sock_accept(struct Fd *fd, struct sockaddr *addr, socklen_t *addrlen)
     return fd2num(nfd);
 }
 
-static int
+static ssize_t
 sock_write(struct Fd *fd, const void *buf, size_t count, off_t offset)
 {
     if (count > 1024)
@@ -147,7 +147,7 @@ sock_write(struct Fd *fd, const void *buf, size_t count, off_t offset)
     return netd_call(&a);
 }
 
-static int
+static ssize_t
 sock_read(struct Fd *fd, void *buf, size_t count, off_t offset)
 {
     if (count > 1024)
